Update asset packing script

+++ AssetPacker.csx
@@ -11,19 +11,32 @@ string outputDir = "assets";
 string manifest = "files.txt";
 string datadir = "data";
 
-int i = 0;
-object s = new object();
+string toRelativeDirStr(string path){
+    string prefix = Path.GetFullPath(Environment.CurrentDirectory);
+    if(!prefix.EndsWith(Path.DirectorySeparatorChar)){prefix += Path.DirectorySeparatorChar;}
+    path = Path.GetFullPath(path);
+    if(path.Length > prefix.Length){
+        path = path.Substring(prefix.Length);
+    }
+    return path;
+}
+
+SHA1 sha1 = SHA1.Create();
+
 string getHash(string file)
 {
-    lock (s)
-    {
-        return (i++).ToString("00000");
-    }
+    string filetime = new FileInfo(file).LastWriteTime.ToString("yyyy\\-MM\\-dd HH\\:mm\\:ss\\.ffffff");
+    string filename = file;
+    string hashstr = filename + filetime;
+    return Convert.ToHexString(sha1.ComputeHash(Encoding.UTF8.GetBytes(hashstr))).ToLower().Replace(" ", "").Replace("-", "").Replace(":", "");
 }
 
 void copyNoOverride(String src, String dest)
 {
-
+    if (File.Exists(dest)){
+        return;
+    }
+    Console.WriteLine(toRelativeDirStr(src)+" => "+toRelativeDirStr(dest));
     File.Copy(src, dest, true);
 }
 
@@ -56,12 +69,23 @@ if (!Directory.Exists(Path.Combine(outputDir, datadir)))
 
 List<string> fileLists = new List<string>();
 
+List<string> tobeDeleted = new List<string>();
+tobeDeleted.AddRange(Directory.EnumerateFiles(Path.Combine(outputDir, datadir)).Select(Path.GetFileName)!);
+
 rescueFiles(inputDir, inputDir, (path, file) =>
 {
     string hash = getHash(file) + Path.GetExtension(file);
     fileLists.Add(hash + " " + path);
+    tobeDeleted.Remove(hash);
     copyNoOverride(file, Path.Combine(outputDir, datadir, hash));
-    Console.WriteLine(path);
+    // Console.WriteLine(path);
 });
 File.WriteAllLines(Path.Combine(outputDir, manifest), fileLists.ToArray());
+
+tobeDeleted.ForEach(f =>{
+    var path = Path.Combine(outputDir, datadir,f);
+    Console.WriteLine("Deleted: "+toRelativeDirStr(path));
+    File.Delete(path);
+});
+
 Console.WriteLine("Done");
